backend/pyproject.toml
[project]
name = "jamtheory-backend"
version = "0.1.0"
description = "JamTheory backend — Cloud Functions for YouTube → Gemini masterclass generation and lick regeneration."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "firebase-admin>=7.3.0",
    "functions-framework>=3.8.2",
    "google-cloud-aiplatform>=1.142.0",
    "google-genai>=1.0.0",
    "google-cloud-firestore>=2.25.0",
    "google-cloud-logging>=3.0.0",
    "google-api-python-client>=2.140.0",
    "pydantic>=2.12.5",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
